Refactor PayPal and email utilities into modules
