+++ website/utils/email.py
@@ -1,0 +1,41 @@
+import os
+import smtplib
+import ssl
+from email.mime.text import MIMEText
+from email.mime.multipart import MIMEMultipart
+from flask import current_app
+
+
+def _env_int(name: str, default: int) -> int:
+    value = os.getenv(name)
+    try:
+        return int(value) if value not in (None, "") else default
+    except ValueError:
+        return default
+
+
+ADMIN_EMAIL = os.getenv("ADMIN_EMAIL", "")
+SMTP_HOST = os.getenv("SMTP_HOST", "")
+SMTP_PORT = _env_int("SMTP_PORT", 587)
+SMTP_USER = os.getenv("SMTP_USER", "")
+SMTP_PASS = os.getenv("SMTP_PASS", "")
+
+
+def send_admin_email(subject: str, html_body: str) -> None:
+    """Send an email to the configured administrator address."""
+    if not (SMTP_HOST and SMTP_USER and SMTP_PASS and ADMIN_EMAIL):
+        current_app.logger.warning(
+            "SMTP not configured; skipping admin email: %s", subject
+        )
+        return
+
+    msg = MIMEMultipart()
+    msg["From"] = SMTP_USER
+    msg["To"] = ADMIN_EMAIL
+    msg["Subject"] = subject
+    msg.attach(MIMEText(html_body, "html"))
+
+    context = ssl.create_default_context()
+    with smtplib.SMTP_SSL(SMTP_HOST, SMTP_PORT, context=context) as server:
+        server.login(SMTP_USER, SMTP_PASS)
+        server.sendmail(SMTP_USER, ADMIN_EMAIL, msg.as_string())
